requirements.txt
fastapi
uvicorn[standard]
onnxruntime
numpy
pillow
python-multipart
httpx